requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
discord==2.3.2
discord.py==2.5.2
ffmpeg-python==0.2.0
filetype==1.2.0
frozenlist==1.5.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.31.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.2.1
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
yt-dlp==2025.2.19